profile-v1/requirements.txt
django==3.0.5
djangorestframework==3.10.3